requirements.txt
Flask>=2.0
pyserial>=3.5
pycountry>=24.6.1
requests>=2.0
